ranslation/requirements.txt
# Core dependencies
requests>=2.31.0
pydantic>=2.0.0
pyyaml>=6.0
pathlib2>=2.3.7; python_version < "3.4"

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0

# Logging and utilities
colorlog>=6.7.0
tqdm>=4.65.0

# Optional LLM API clients
openai>=1.0.0
anthropic>=0.7.0
